requirements.txt
alembic==1.9.0
annotated-types==0.7.0
anyio==4.7.0
asyncpg==0.30.0
attrs==24.3.0
bcrypt==4.0.1
certifi==2024.12.14
cfgv==3.4.0
click==8.1.8
colorama==0.4.6
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
envparse==0.2.0
fastapi==0.115.6
filelock==3.16.1
greenlet==3.1.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
identify==2.6.4
idna==3.10
iniconfig==2.0.0
Mako==1.3.8
MarkupSafe==3.0.2
nodeenv==1.9.1
packaging==24.2
passlib==1.7.4
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==2.21.0
prometheus_client==0.21.1
psycopg2==2.9.10
psycopg2-binary==2.9.5
pyasn1==0.6.1
pydantic==2.10.4
pydantic-settings
pydantic_core==2.27.2
pytest==7.2.0
pytest-asyncio==0.20.3
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0.2
rfc3986==1.5.0
rsa==4.9
sentry-sdk==2.19.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
starlette-exporter==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.28.0
